[FE] refactor: 리덕스 초기 세팅
