+++ client/src/_reducers/users_reducer.js
@@ -1,4 +1,4 @@
-import { GET_USERS } from '../_actions/user_action';
+import { GET_USERS } from '../_actions/users_action';
 
 export default function users(state = {}, action) {
   switch (action.type) {
