+++ client/src/api/requests.js
@@ -1,1 +1,5 @@
 export const USERS_URL = '/api/users';
+export const QUESTIONS_URL = '/api/questions';
+export const ANSWER_URL = '/api/answer';
+export const SEARCH_URL = '/api/search';
+export const TAGS_URL = '/api/tags';
